+++ src/xpk/utils.py
@@ -19,6 +19,8 @@ import os
 import re
 import sys
 import tempfile
+import urllib.request
+from urllib.error import ContentTooShortError
 
 
 def chunks(lst: list, n: int):
@@ -165,3 +167,15 @@ def directory_path_type(value):
         f'Directory path is invalid. User provided path was {value}'
     )
   return value
+
+
+def download_files_from_github_into_dir(
+    path: str, urls: list[tuple[str, str]]
+) -> None:
+  for url, fn in urls:
+    target = os.path.join(path, fn)
+    try:
+      urllib.request.urlretrieve(url, target)
+    except ContentTooShortError as e:
+      xpk_print(f'downloading kjob CRD {fn} failed due to {e.content}')
+      xpk_exit(1)
